+++ src/Patches/GeneralManagerPatch.cs
@@ -1,5 +1,6 @@
 using System;
 using HarmonyLib;
+using Il2Cpp;
 using JetBrains.Annotations;
 using UnityEngine;