+++ src/Patches/MainMenuPatch.cs
@@ -1,5 +1,6 @@
 using System;
 using HarmonyLib;
+using Il2Cpp;
 using UnityEngine;
 
 namespace PPDSAP.Patches;