+++ Changelog.md
@@ -1,3 +1,9 @@
+# v0.1.2
+
+- [Client] Ported mod from BepinEx modloader to Melon Loader
+- [Client] The duck's names in the pool will be random names from the multiworld
+
+---
 # v.0.1.1-hf1
 
 - ~~added a single `?`~~ Fixed a bug with having no save to resume from causing the mod to panic